Use knex for query generation
This switches from pgpool to knex for query generation.  For the most
part, the translation is simple, but the handling of transactions for
cloudwatch events needed some refactoring: it now operates within a
`modifyInstance` method which takes a callback.  When that callback
exits, successfully or not, its transaction is completed (COMMIT or
ROLLBACK).  This ensures that all code-paths correctly handle the
transactions.

Conditional queries now use `in` in place of `x = a OR x = b OR ..`.